excise db func of connect and get
